+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/EmptyState/ODSEmptyStateView.swift
@@ -19,7 +19,7 @@ import SwiftUI
 /// A button can also be added.
 ///
 public struct ODSEmptyStateView: View {
-
+    
     // =======================
     // MARK: Stored properties
     // =======================
@@ -81,6 +81,26 @@ public struct ODSEmptyStateView: View {
             }
             .padding(.horizontal, ODSSpacing.m)
             .padding(.vertical, ODSSpacing.s)
+        }
+    }
+    
+    public enum DefaultAssets {
+        case uerCleared
+        case error
+        case firstUse
+        case noData
+        
+        var name: String {
+            switch self {
+            case .uerCleared:
+                return "il_emptyStateUserCleared"
+            case .error:
+                return "il_emptyStateError"
+            case .firstUse:
+                return "il_emptyStateFirstUse"
+            case .noData:
+                return "il_emptyStateNoData"
+            }
         }
     }
 }