+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/MoreApps/Data/LocalAppsPlusRepository.swift
@@ -36,14 +36,14 @@ struct LocalAppsPlusRepository: MoreAppsRepositoryProtocol {
     // MARK: MoreAppsRepositoryProtocol - Impl
     // =======================================
 
-    func availableAppsList() async throws -> MoreAppsAppsList {
+    func availableAppsList() async throws -> MoreAppsList {
         do {
             let rawData = try Data(contentsOf: feedURL)
             let appsPlusAppsList: AppsPlusListDTO = try JSONDecoder().decode(AppsPlusDTO.self, from: rawData).items[0]
             let mapper = AppsPlusToMoreAppsMapper()
             let apps = mapper.appsDetails(from: appsPlusAppsList)
             let sections = mapper.appsSections(from: appsPlusAppsList)
-            return MoreAppsAppsList(sections: sections, apps: apps)
+            return MoreAppsList(sections: sections, apps: apps)
         } catch {
             ODSLogger.error("(ノಠ益ಠ)ノ彡┻━┻ Failed to decode local AppsPlus file: '\(error.localizedDescription)'")
             throw MoreAppsServiceErrors.jsonDecodingFailure