+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [DemoApp/SDK] Rename recirculation module to MoreApps, and use EmptyState Assets from Theme bundle ([#801](https://github.com/Orange-OpenSource/ods-ios/issues/801))
+- [DemoApp/SDK] Rename recirculatin module to MoreApps ([#801](https://github.com/Orange-OpenSource/ods-ios/issues/801))
 - [Doc] Update components and modules documentation to add a description of the api ([#759](https://github.com/Orange-OpenSource/ods-ios/issues/759))
 - [Tooling] Update dependency SwiftFormat/CLI to v0.53.5
 - [DemoApp] The main color names, visible on the screen are not vocalized ([#720](https://github.com/Orange-OpenSource/ods-ios/issues/720))